TASK: Refactor several protected methods from NodeDataRepository to public methods in NodeDataUtility and make createQueryBuilder public
- withoutRemovedNodes (move to NodeDataUtility)
- addDimensionJoinConstraintsToQueryBuilder (move to NodeDataUtility)
- reduceNodeVariantsByWorkspacesAndDimensions (move to NodeDataUtility)
- collectWorkspaceAndAllBaseWorkspaces (move to NodeDataUtility)
- createQueryBuilder (make public)
