+++ Neos.ContentRepository/Classes/Domain/Repository/NodeDataRepository.php
@@ -21,6 +21,7 @@ use Neos\ContentRepository\Domain\Repository\Query\NodeParentFilter;
 use Neos\ContentRepository\Domain\Repository\Query\NodePropertyFilter;
 use Neos\ContentRepository\Domain\Repository\Query\NodePropertyFulltextFilter;
 use Neos\ContentRepository\Domain\Repository\Query\NodeTypeFilter;
+use Neos\ContentRepository\Domain\Utility\NodeDataUtility;
 use Neos\Flow\Annotations as Flow;
 use Neos\Flow\Log\Utility\LogEnvironment;
 use Neos\Flow\Persistence\QueryInterface;
@@ -194,16 +195,16 @@ class NodeDataRepository extends Repository
             return $workspace->getRootNodeData();
         }
 
-        $workspaces = $this->collectWorkspaceAndAllBaseWorkspaces($workspace);
+        $workspaces = NodeDataUtility::collectWorkspaceAndAllBaseWorkspaces($workspace);
         $nodes = $this->findRawNodesByPath($path, $workspace, $dimensions);
         $dimensions = $dimensions === null ? [] : $dimensions;
-        $foundNodes = $this->reduceNodeVariantsByWorkspacesAndDimensions($nodes, $workspaces, $dimensions);
+        $foundNodes = NodeDataUtility::reduceNodeVariantsByWorkspacesAndDimensions($nodes, $workspaces, $dimensions);
         $foundNodes = $this->filterNodeDataByBestMatchInContext($foundNodes, $workspace, $dimensions, $removedNodes);
 
         if ($removedNodes === true) {
             $foundNodes = $this->onlyRemovedNodes($foundNodes);
         } elseif ($removedNodes === false) {
-            $foundNodes = $this->withoutRemovedNodes($foundNodes);
+            $foundNodes = NodeDataUtility::withoutRemovedNodes($foundNodes);
         }
 
         if ($foundNodes !== []) {
@@ -223,10 +224,10 @@ class NodeDataRepository extends Repository
      */
     public function findShadowNodeByPath($path, Workspace $workspace, array $dimensions = null)
     {
-        $workspaces = $this->collectWorkspaceAndAllBaseWorkspaces($workspace);
+        $workspaces = NodeDataUtility::collectWorkspaceAndAllBaseWorkspaces($workspace);
         $nodes = $this->findRawNodesByPath($path, $workspace, $dimensions, true);
         $dimensions = $dimensions === null ? [] : $dimensions;
-        $foundNodes = $this->reduceNodeVariantsByWorkspacesAndDimensions($nodes, $workspaces, $dimensions);
+        $foundNodes = NodeDataUtility::reduceNodeVariantsByWorkspacesAndDimensions($nodes, $workspaces, $dimensions);
         $foundNodes = $this->onlyRemovedNodes($foundNodes);
 
         if ($foundNodes !== []) {
@@ -276,7 +277,7 @@ class NodeDataRepository extends Repository
         }
         $queryBuilder = $this->createQueryBuilder($workspaces);
         if ($dimensions !== null) {
-            $this->addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
+            NodeDataUtility::addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
         }
         $this->addPathConstraintToQueryBuilder($queryBuilder, $path);
         if ($onlyShadowNodes) {
@@ -357,7 +358,7 @@ class NodeDataRepository extends Repository
             $queryBuilder->andWhere('n.movedTo IS NULL');
         }
         if ($dimensions !== null) {
-            $this->addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
+            NodeDataUtility::addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
         } else {
             $dimensions = [];
         }
@@ -366,9 +367,9 @@ class NodeDataRepository extends Repository
         $query = $queryBuilder->getQuery();
         $nodes = $query->getResult();
 
-        $foundNodes = $this->reduceNodeVariantsByWorkspacesAndDimensions($nodes, $workspaces, $dimensions);
+        $foundNodes = NodeDataUtility::reduceNodeVariantsByWorkspacesAndDimensions($nodes, $workspaces, $dimensions);
         if ($removedNodes === false) {
-            $foundNodes = $this->withoutRemovedNodes($foundNodes);
+            $foundNodes = NodeDataUtility::withoutRemovedNodes($foundNodes);
         }
 
         if ($foundNodes !== []) {
@@ -594,11 +595,11 @@ class NodeDataRepository extends Repository
      */
     protected function getNodeDataForParentAndNodeType($parentPath, $nodeTypeFilter, Workspace $workspace, array $dimensions = null, $removedNodes = false, $recursive = false)
     {
-        $workspaces = $this->collectWorkspaceAndAllBaseWorkspaces($workspace);
+        $workspaces = NodeDataUtility::collectWorkspaceAndAllBaseWorkspaces($workspace);
 
         $queryBuilder = $this->createQueryBuilder($workspaces);
         if ($dimensions !== null) {
-            $this->addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
+            NodeDataUtility::addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
         } else {
             $dimensions = [];
         }
@@ -610,13 +611,13 @@ class NodeDataRepository extends Repository
         $query = $queryBuilder->getQuery();
         $nodes = $query->getResult();
 
-        $foundNodes = $this->reduceNodeVariantsByWorkspacesAndDimensions($nodes, $workspaces, $dimensions);
+        $foundNodes = NodeDataUtility::reduceNodeVariantsByWorkspacesAndDimensions($nodes, $workspaces, $dimensions);
         $foundNodes = $this->filterNodeDataByBestMatchInContext($foundNodes, $workspaces[0], $dimensions, $removedNodes);
 
         if ($removedNodes === true) {
             $foundNodes = $this->onlyRemovedNodes($foundNodes);
         } elseif ($removedNodes === false) {
-            $foundNodes = $this->withoutRemovedNodes($foundNodes);
+            $foundNodes = NodeDataUtility::withoutRemovedNodes($foundNodes);
         }
 
         return $foundNodes;
@@ -634,7 +635,7 @@ class NodeDataRepository extends Repository
     public function findByParentWithoutReduce($parentPath, Workspace $workspace)
     {
         $parentPath = strtolower($parentPath);
-        $workspaces = $this->collectWorkspaceAndAllBaseWorkspaces($workspace);
+        $workspaces = NodeDataUtility::collectWorkspaceAndAllBaseWorkspaces($workspace);
 
         $queryBuilder = $this->createQueryBuilder($workspaces);
         $this->addParentPathConstraintToQueryBuilder($queryBuilder, $parentPath);
@@ -672,7 +673,7 @@ class NodeDataRepository extends Repository
      */
     public function findByIdentifierWithoutReduce($identifier, Workspace $workspace)
     {
-        $workspaces = $this->collectWorkspaceAndAllBaseWorkspaces($workspace);
+        $workspaces = NodeDataUtility::collectWorkspaceAndAllBaseWorkspaces($workspace);
 
         $queryBuilder = $this->createQueryBuilder($workspaces);
         $this->addIdentifierConstraintToQueryBuilder($queryBuilder, $identifier);
@@ -972,12 +973,12 @@ class NodeDataRepository extends Repository
             throw new \InvalidArgumentException('Invalid paths: path of starting point must be first part of end point path.', 1284391181);
         }
 
-        $workspaces = $this->collectWorkspaceAndAllBaseWorkspaces($workspace);
+        $workspaces = NodeDataUtility::collectWorkspaceAndAllBaseWorkspaces($workspace);
 
         $queryBuilder = $this->createQueryBuilder($workspaces);
 
         if ($dimensions !== null) {
-            $this->addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
+            NodeDataUtility::addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
         } else {
             $dimensions = [];
         }
@@ -1001,11 +1002,11 @@ class NodeDataRepository extends Repository
 
         $query = $queryBuilder->getQuery();
         $foundNodes = $query->getResult();
-        $foundNodes = $this->reduceNodeVariantsByWorkspacesAndDimensions($foundNodes, $workspaces, $dimensions);
+        $foundNodes = NodeDataUtility::reduceNodeVariantsByWorkspacesAndDimensions($foundNodes, $workspaces, $dimensions);
         $foundNodes = $this->filterNodeDataByBestMatchInContext($foundNodes, $workspaces[0], $dimensions, $includeRemovedNodes);
 
         if ($includeRemovedNodes === false) {
-            $foundNodes = $this->withoutRemovedNodes($foundNodes);
+            $foundNodes = NodeDataUtility::withoutRemovedNodes($foundNodes);
         }
 
         $nodesByDepth = [];
@@ -1192,19 +1193,6 @@ class NodeDataRepository extends Repository
     }
 
     /**
-     * Returns a subset of $nodes which are not flagged as removed.
-     *
-     * @param array $nodes NodeData including removed entries
-     * @return array Only those NodeData instances which are not flagged as removed
-     */
-    public function withoutRemovedNodes(array $nodes)
-    {
-        return array_filter($nodes, function (NodeData $node) {
-            return !$node->isRemoved();
-        });
-    }
-
-    /**
      * Returns a subset of $nodes which are flagged as removed.
      *
      * @param array $nodes NodeData including removed entries
@@ -1241,92 +1229,6 @@ class NodeDataRepository extends Repository
      */
     protected function emitRepositoryObjectsPersisted()
     {
-    }
-
-    /**
-     * If $dimensions is not empty, adds join constraints to the given $queryBuilder
-     * limiting the query result to matching hits.
-     *
-     * @param QueryBuilder $queryBuilder
-     * @param array $dimensions
-     * @return void
-     */
-    public function addDimensionJoinConstraintsToQueryBuilder(QueryBuilder $queryBuilder, array $dimensions)
-    {
-        $count = 0;
-        foreach ($dimensions as $dimensionName => $dimensionValues) {
-            $dimensionAlias = 'd' . $count;
-            $queryBuilder->andWhere(
-                'EXISTS (SELECT ' . $dimensionAlias . ' FROM Neos\ContentRepository\Domain\Model\NodeDimension ' . $dimensionAlias . ' WHERE ' . $dimensionAlias . '.nodeData = n AND ' . $dimensionAlias . '.name = \'' . $dimensionName . '\' AND ' . $dimensionAlias . '.value IN (:' . $dimensionAlias . ')) ' .
-                'OR NOT EXISTS (SELECT ' . $dimensionAlias . '_c FROM Neos\ContentRepository\Domain\Model\NodeDimension ' . $dimensionAlias . '_c WHERE ' . $dimensionAlias . '_c.nodeData = n AND ' . $dimensionAlias . '_c.name = \'' . $dimensionName . '\')'
-            );
-            $queryBuilder->setParameter($dimensionAlias, $dimensionValues);
-            $count++;
-        }
-    }
-
-    /**
-     * Given an array with duplicate nodes (from different workspaces and dimensions) those are reduced to uniqueness (by node identifier)
-     *
-     * @param array $nodes NodeData result with multiple and duplicate identifiers (different nodes and redundant results for node variants with different dimensions)
-     * @param array $workspaces
-     * @param array $dimensions
-     * @return array Array of unique node results indexed by identifier
-     * @throws Exception\NodeException
-     */
-    public function reduceNodeVariantsByWorkspacesAndDimensions(array $nodes, array $workspaces, array $dimensions)
-    {
-        $reducedNodes = [];
-
-        $minimalDimensionPositionsByIdentifier = [];
-        foreach ($nodes as $node) {
-            /** @var NodeData $node */
-            $nodeDimensions = $node->getDimensionValues();
-
-            // Find the position of the workspace, a smaller value means more priority
-            $workspaceNames = array_map(
-                function (Workspace $workspace) {
-                    return $workspace->getName();
-                },
-                $workspaces
-            );
-            $workspacePosition = array_search($node->getWorkspace()->getName(), $workspaceNames);
-            if ($workspacePosition === false) {
-                throw new Exception\NodeException(sprintf('Node workspace "%s" not found in allowed workspaces (%s), this could result from a detached workspace entity in the context.', $node->getWorkspace()->getName(), implode($workspaceNames, ', ')), 1413902143);
-            }
-
-            // Find positions in dimensions, add workspace in front for highest priority
-            $dimensionPositions = [];
-
-            // Special case for no dimensions
-            if ($dimensions === []) {
-                // We can just decide if the given node has no dimensions.
-                $dimensionPositions[] = ($nodeDimensions === []) ? 0 : 1;
-            }
-
-            foreach ($dimensions as $dimensionName => $dimensionValues) {
-                if (isset($nodeDimensions[$dimensionName])) {
-                    foreach ($nodeDimensions[$dimensionName] as $nodeDimensionValue) {
-                        $position = array_search($nodeDimensionValue, $dimensionValues);
-                        $dimensionPositions[$dimensionName] = isset($dimensionPositions[$dimensionName]) ? min($dimensionPositions[$dimensionName],
-                            $position) : $position;
-                    }
-                } else {
-                    $dimensionPositions[$dimensionName] = isset($dimensionPositions[$dimensionName]) ? min($dimensionPositions[$dimensionName],
-                            PHP_INT_MAX) : PHP_INT_MAX;
-                }
-            }
-            $dimensionPositions[] = $workspacePosition;
-
-            $identifier = $node->getIdentifier();
-            // Yes, it seems to work comparing arrays that way!
-            if (!isset($minimalDimensionPositionsByIdentifier[$identifier]) || $dimensionPositions < $minimalDimensionPositionsByIdentifier[$identifier]) {
-                $reducedNodes[$identifier] = $node;
-                $minimalDimensionPositionsByIdentifier[$identifier] = $dimensionPositions;
-            }
-        }
-
-        return $reducedNodes;
     }
 
     /**
@@ -1428,7 +1330,7 @@ class NodeDataRepository extends Repository
     public function findByPathWithoutReduce($path, Workspace $workspace, $includeRemovedNodes = false, $recursive = false)
     {
         $path = strtolower($path);
-        $workspaces = $this->collectWorkspaceAndAllBaseWorkspaces($workspace);
+        $workspaces = NodeDataUtility::collectWorkspaceAndAllBaseWorkspaces($workspace);
 
         $queryBuilder = $this->createQueryBuilder($workspaces);
         $this->addPathConstraintToQueryBuilder($queryBuilder, $path, $recursive);
@@ -1444,7 +1346,7 @@ class NodeDataRepository extends Repository
 
         $foundNodes = $this->reduceNodeVariantsByWorkspaces($foundNodes, $workspaces);
         if ($includeRemovedNodes === false) {
-            $foundNodes = $this->withoutRemovedNodes($foundNodes);
+            $foundNodes = NodeDataUtility::withoutRemovedNodes($foundNodes);
         }
 
         return $foundNodes;
@@ -1518,7 +1420,7 @@ class NodeDataRepository extends Repository
      * @param array $workspaces
      * @return QueryBuilder
      */
-    public function createQueryBuilder(array $workspaces)
+    public function createQueryBuilder(array $workspaces): QueryBuilder
     {
         /** @var QueryBuilder $queryBuilder */
         $queryBuilder = $this->entityManager->createQueryBuilder();
@@ -1593,7 +1495,7 @@ class NodeDataRepository extends Repository
      */
     protected function filterNodeDataByBestMatchInContext(array $nodeDataObjects, Workspace $workspace, array $dimensions, $includeRemovedNodes = false)
     {
-        $workspaces = $this->collectWorkspaceAndAllBaseWorkspaces($workspace);
+        $workspaces = NodeDataUtility::collectWorkspaceAndAllBaseWorkspaces($workspace);
         $nonPersistedNodes = [];
         $nodeIdentifier = [];
 
@@ -1614,7 +1516,7 @@ class NodeDataRepository extends Repository
 
         $queryBuilder = $this->createQueryBuilder($workspaces);
         if ($dimensions !== null) {
-            $this->addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
+            NodeDataUtility::addDimensionJoinConstraintsToQueryBuilder($queryBuilder, $dimensions);
         } else {
             $dimensions = [];
         }
@@ -1628,35 +1530,18 @@ class NodeDataRepository extends Repository
         $query = $queryBuilder->getQuery();
         $nodes = $query->getResult();
         $foundNodes = array_merge($nodes, $nonPersistedNodes);
-        $foundNodes = $this->reduceNodeVariantsByWorkspacesAndDimensions($foundNodes, $workspaces, $dimensions);
+        $foundNodes = NodeDataUtility::reduceNodeVariantsByWorkspacesAndDimensions($foundNodes, $workspaces, $dimensions);
 
         if ($includeRemovedNodes === true) {
             $foundNodes = $this->onlyRemovedNodes($foundNodes);
         } elseif ($includeRemovedNodes === false) {
-            $foundNodes = $this->withoutRemovedNodes($foundNodes);
+            $foundNodes = NodeDataUtility::withoutRemovedNodes($foundNodes);
         }
 
         /** @var NodeData $nodeData */
         return array_filter($nodeDataObjects, function (NodeData $nodeData) use ($foundNodes) {
             return (isset($foundNodes[$nodeData->getIdentifier()]) && $foundNodes[$nodeData->getIdentifier()] === $nodeData);
         });
-    }
-
-    /**
-     * Returns an array that contains the given workspace and all base (parent) workspaces of it.
-     *
-     * @param Workspace $workspace
-     * @return array
-     */
-    public function collectWorkspaceAndAllBaseWorkspaces(Workspace $workspace)
-    {
-        $workspaces = [];
-        while ($workspace !== null) {
-            $workspaces[] = $workspace;
-            $workspace = $workspace->getBaseWorkspace();
-        }
-
-        return $workspaces;
     }
 
     /**
